Rename pyro.contrib.outbreak -> pyro.contrib.epidemiology
