+++ docs/source/contrib.outbreak.rst
@@ -1,11 +1,11 @@
 Epidemiological Models
 ======================
-.. automodule:: pyro.contrib.outbreak
+.. automodule:: pyro.contrib.epidemiology
 
 .. warning:: Code in ``pyro.contrib.forecast`` is under development.
     This code makes no guarantee about maintaining backwards compatibility.
 
-``pyro.contrib.outbreak`` is a framework for experimenting with a restricted
+``pyro.contrib.epidemiology`` is a framework for experimenting with a restricted
 class of stochastic discrete-time discrete-count compartmental models. This
 framework implements **inference** via Hamiltonian Monte Carlo, **prediction**
 of latent variables, and **forecasting** future trajectories.
@@ -15,12 +15,12 @@ For explanation of the underlying machinery in this framework, see the
 
 Base Compartmental Model
 ------------------------
-.. automodule:: pyro.contrib.outbreak.compartmental
+.. automodule:: pyro.contrib.epidemiology.compartmental
     :members:
     :member-order: bysource
 
 SIR Models
 ----------
-.. automodule:: pyro.contrib.outbreak.sir
+.. automodule:: pyro.contrib.epidemiology.sir
     :members:
     :member-order: bysource
