+++ docs/source/index.rst
@@ -33,12 +33,12 @@ Pyro Documentation
    contrib.bnn
    contrib.cevae
    contrib.easyguide
+   contrib.epidemiology
    contrib.examples
    contrib.forecast
    contrib.gp
    contrib.minipyro
    contrib.oed
-   contrib.outbreak
    contrib.timeseries
    contrib.tracking
