+++ pyro/contrib/__init__.py
@@ -9,15 +9,15 @@ Contributed Code
     This code makes no guarantee about maintaining backwards compatibility.
 """
 
-from pyro.contrib import autoname, bnn, easyguide, forecast, gp, oed, outbreak, tracking
+from pyro.contrib import autoname, bnn, easyguide, epidemiology, forecast, gp, oed, tracking
 
 __all__ = [
     "autoname",
     "bnn",
     "easyguide",
+    "epidemiology",
     "forecast",
     "gp",
     "oed",
-    "outbreak",
     "tracking",
 ]
