+++ pyro/contrib/epidemiology/compartmental.py
