+++ pyro/contrib/epidemiology/sir.py
