+++ pyro/contrib/epidemiology/util.py
